Dual distribution: npm package + asset pipeline
