--- app/assets/javascripts/turbo/index.js
@@ -1,3 +1,0 @@
-//= link ./streams
-//= link ./action_cable
-//= require ./turbo@7.0.0-beta1
