+++ app/helpers/turbo/includes_helper.rb
@@ -1,5 +1,5 @@
 module Turbo::IncludesHelper
   def turbo_include_tags
-    javascript_include_tag("turbo", type: "module") + javascript_include_tag("turbo/streams", type: "module", "data-action-cable-src": asset_path("turbo/action_cable"))
+    javascript_include_tag("turbo", type: "module")
   end
 end
