+++ app/helpers/turbo/streams_helper.rb
@@ -17,6 +17,6 @@ module Turbo::StreamsHelper
   # (when Current.account.id = 5). Updates to this stream can be sent like
   # <tt>entry.broadcast_append_to entry.account, :entries, contrainer: "entries"</tt>.
   def turbo_stream_from(*streamables)
-    tag.turbo_stream_from channel: "Turbo::StreamsChannel", "signed-stream-name": Turbo::StreamsChannel.signed_stream_name(streamables)
+    tag.turbo_cable_stream_source channel: "Turbo::StreamsChannel", "signed-stream-name": Turbo::StreamsChannel.signed_stream_name(streamables)
   end
 end
