+++ app/javascript/turbo/index.js
@@ -1,0 +1,3 @@
+import "./cable_stream_source_element"
+export * as Turbo from "@hotwired/turbo"
+export * as cable from "./cable"
